+++ common/src/main/java/io/github/mortuusars/envelope/world/item/crafting/mail/MailLetterBroadcastingRecipe.java
@@ -29,6 +29,7 @@ import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 
 import java.util.Map;
+import java.util.Optional;
 import java.util.function.Function;
 
 public class MailLetterBroadcastingRecipe extends CustomMailRecipe {
@@ -37,6 +38,7 @@ public class MailLetterBroadcastingRecipe extends CustomMailRecipe {
     private final ItemStack result = Mail.createLetter(Component.empty())
           .itemName(Component.translatable("letter.envelope.broadcast_report.name"))
           .get();
+    private final Optional<Component> info = Optional.of(Component.translatable("recipe.envelope.mailing.mail_service.letter_broadcasting.info"));
 
     public MailLetterBroadcastingRecipe(ServiceAddress address, NonNullList<Ingredient> ingredients) {
         super(address);
@@ -56,6 +58,11 @@ public class MailLetterBroadcastingRecipe extends CustomMailRecipe {
     @Override
     public boolean isOneCraftPerDelivery() {
         return true; // Prevent spamming and spawning tons of couriers
+    }
+
+    @Override
+    public Optional<Component> getInfo() {
+        return info;
     }
 
     @Override